[Fix] Issue processing `sigma` in the forward of GaussianNoise
During computation of the forward, it was required to transform the
shape of sigma:

type(sigma) = torch.Tensor
sigma.shape = (bs,)
->
type(sigma) = torch.Tensor
sigma.shape = (bs,1,1,...)

Sigma is supposed to be stored as an attribute of GaussianNoise.
However it is also possible to be given a new sigma directly in the
forward.

The processing was done only when sigma was passed as an argument in
the forward which causes error when doing the forward.