make SeparablePrior work with TVPrior
The previous implementation of SeparablePrior fails when using TVPrior
as the base prior. This is because the input components have one of
their dimensions squeezed out before being passed to the proximity
operator of the base prior. Moreover, unlike L1Prior, the one used in
the tests, TVPrior does not support input tensors of shape (B, H, W).

Even though total variation priors are generally not separable, we still
view this as a bug worth fixing as it might impact other more natural
priors.

We fix the bug by using `torch.split` instead of `torch.unbind` as
suggested by @gRox167.

We also make the loop cleaner by using `zip` instead of `enumerate`.